(enhance #6) shmqlib 포팅
1) HTTPC/S <--> AHIF 간 UDM/R 과 동일한 shmqlib 사용토록 반영
2) ~/src/HTTP/MakeVar 에서 target project 선택하여 선택 빌드 되도록 반영
3) Make -Wall 옵션 적용하여 미사용 변수와 일부 워닝 제거

+++ httpc/command.c
@@ -48,7 +48,7 @@ void message_handle(evutil_socket_t fd, short what, void *arg)
 				continue;
 #endif
 			default:
-				APPLOG(APPLOG_ERR, "not yet ready (mtype:%d)", msg->mtype);
+				APPLOG(APPLOG_ERR, "not yet ready (mtype:%ld)", (long)msg->mtype); // fuck! genq_mtype_t, it just long
 				continue;
 		}
 	}
